Move LoadPodFromFile to //pkg/apis/core/v1/helper
LoadPodFromFile currently resides in //pkg/volume/util. However, it is not
volume specific function and some third parties (most notably kubeadm) are
importing the whole of //pkg/volume/util just to use it. This causes a lot
of unneeded dependencies to be included.
The solution is to move LoadPodFromFile to a more appropriate location and
//pkg/apis/core/v1/helper is the most obvious one.
