+++ cmd/kube-controller-manager/app/BUILD
@@ -32,6 +32,7 @@ go_library(
         "//pkg/apis/certificates/install:go_default_library",
         "//pkg/apis/componentconfig:go_default_library",
         "//pkg/apis/core/install:go_default_library",
+        "//pkg/apis/core/v1/helper:go_default_library",
         "//pkg/apis/events/install:go_default_library",
         "//pkg/apis/extensions/install:go_default_library",
         "//pkg/apis/policy/install:go_default_library",
@@ -104,7 +105,6 @@ go_library(
         "//pkg/volume/rbd:go_default_library",
         "//pkg/volume/scaleio:go_default_library",
         "//pkg/volume/storageos:go_default_library",
-        "//pkg/volume/util:go_default_library",
         "//pkg/volume/vsphere_volume:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
