+++ pkg/apis/core/v1/helper/helpers_test.go
@@ -18,6 +18,8 @@ package helper
 
 import (
 	"fmt"
+	"io/ioutil"
+	"os"
 	"reflect"
 	"testing"
 
@@ -790,5 +792,89 @@ func TestMatchNodeSelectorTerms(t *testing.T) {
 				t.Errorf("MatchNodeSelectorTermsORed() = %v, want %v", got, tt.want)
 			}
 		})
+	}
+}
+
+func TestLoadPodFromFile(t *testing.T) {
+	tests := []struct {
+		name        string
+		content     string
+		expectError bool
+	}{
+		{
+			"yaml",
+			`
+apiVersion: v1
+kind: Pod
+metadata:
+  name: testpod
+spec:
+  containers:
+    - image: k8s.gcr.io/busybox
+`,
+			false,
+		},
+
+		{
+			"json",
+			`
+{
+  "apiVersion": "v1",
+  "kind": "Pod",
+  "metadata": {
+    "name": "testpod"
+  },
+  "spec": {
+    "containers": [
+      {
+        "image": "k8s.gcr.io/busybox"
+      }
+    ]
+  }
+}`,
+			false,
+		},
+
+		{
+			"invalid pod",
+			`
+apiVersion: v1
+kind: Pod
+metadata:
+  name: testpod
+spec:
+  - image: k8s.gcr.io/busybox
+`,
+			true,
+		},
+	}
+
+	for _, test := range tests {
+		tempFile, err := ioutil.TempFile("", "podfile")
+		defer os.Remove(tempFile.Name())
+		if err != nil {
+			t.Fatalf("cannot create temporary file: %v", err)
+		}
+		if _, err = tempFile.Write([]byte(test.content)); err != nil {
+			t.Fatalf("cannot save temporary file: %v", err)
+		}
+		if err = tempFile.Close(); err != nil {
+			t.Fatalf("cannot close temporary file: %v", err)
+		}
+
+		pod, err := LoadPodFromFile(tempFile.Name())
+		if test.expectError {
+			if err == nil {
+				t.Errorf("test %q expected error, got nil", test.name)
+			}
+		} else {
+			// no error expected
+			if err != nil {
+				t.Errorf("error loading pod %q: %v", test.name, err)
+			}
+			if pod == nil {
+				t.Errorf("test %q expected pod, got nil", test.name)
+			}
+		}
 	}
 }
